Bit-for-bit restartability established for mpas-ocean
ocn_comp_nuopc.F90 had its import of coupler fields fixed
and this now establishes bit-for-bit restartability of the
restart files for mpas-ocean. Some modifications to defaults for
diagnostic output were made so that restarted runs produce
bit-for-bit reproducible output as well for some of the
diagnostics. Monthly mean time average output is not yet reproducible
as the first time step after initialization fails to accumulate