Modify the imported swnet check
Fixing bit-for-bit restartability required adding an additional
import from the coupler at initialization. It turns out that
not all fluxes were defined at this point and this tripped
up the check for positive SW in this added call. These changes
allow this check to be skipped for this additional initial import.