Update cvmix to v1.0
mpas-ocean is now using cvmix v1.0, with the Apache License 2.0.

The calls to a couple of the cvmix routines had to be modified as
this version of cvmix changed a scalar argument to a vector. But
these arguments are also now optional. As they are not being
used with default values of cvmix options they are removed from
the calling mpas-ocean routine.